% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_admtab.R
\name{admtab_server}
\alias{admtab_server}
\title{Modul for Administrative tabeller-fane i NRA sin shiny-app på Rapporteket}
\usage{
admtab_server(id, RegData, hvd_session, skjemaoversikt)
}
\value{
Serverdel av modul for adm. tab
}
\description{
Kun til bruk i Shiny
}
