% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_admtab.R
\name{admtab_ui}
\alias{admtab_ui}
\title{Modul for Administrative tabeller-fane i NRA sin shiny-app på Rapporteket}
\usage{
admtab_ui(id)
}
\value{
UI-del av modul for adm. tab
}
\description{
Kun til bruk i Shiny
}
