% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_andel_tid.R
\name{andeler_tid_server}
\alias{andeler_tid_server}
\title{Modul for tidsfigur i NRA sin shiny-app på Rapporteket}
\usage{
andeler_tid_server(id, reshID, RegData, hvd_session, BrValg, userRole)
}
\arguments{
\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}
}
\value{
server-del av tidsfigur
}
\description{
Kun til bruk i Shiny
}
