% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_andel_tid.R
\name{andeler_tid_ui}
\alias{andeler_tid_ui}
\title{Modul for tidsfigur i NRA sin shiny-app på Rapporteket}
\usage{
andeler_tid_ui(id)
}
\value{
ui-del av tidsfigur
}
\description{
Kun til bruk i Shiny
}
