% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomkonf.R
\name{binomkonf}
\alias{binomkonf}
\title{Beregn konfidensintervall for binomisk data med n suksess av N forsøk}
\usage{
binomkonf(n, N, konfnivaa = 0.95)
}
\arguments{
\item{n}{- Antall suksess, skalar eller vektor}

\item{N}{- Antall forsøk, skalar eller vektor av samme lengde som n}

\item{konfnivaa}{- Konfidensnivået til test (Default 0.95)}
}
\value{
Nedre og øvre grense til konfidensintervallet
}
\description{
Denne funksjonen beregner konfidensintervall for binomisk fordelt data
}
