% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrValg.R
\name{BrValg}
\alias{BrValg}
\title{Definer lister for brukerkontroller}
\usage{
BrValg(RegData)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}
}
\value{
BrValg En liste med brukerkontrollvektorer
}
\description{
Denne funksjonen lager navnede vektorer til bruk i brukerkontroller i Shiny
}
