% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_datadump.R
\name{datadump_server}
\alias{datadump_server}
\title{Modul for datadump-fane i NRA sin shiny-app på Rapporteket}
\usage{
datadump_server(id, reshID, userRole, hvd_session)
}
\value{
Datadump av registerets tabeller
}
\description{
Kun til bruk i Shiny
}
