% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_datadump.R
\name{datadump_ui}
\alias{datadump_ui}
\title{Modul for datadump-fane i NRA sin shiny-app på Rapporteket}
\usage{
datadump_ui(id)
}
\value{
Datadump av registerets tabeller
}
\description{
Kun til bruk i Shiny
}
