% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nra_dateInput2.R
\name{dateInput2}
\alias{dateInput2}
\title{R-shiny alternativ til dateInput med mulighet til  bare år, måned og år}
\usage{
dateInput2(inputId, label, minview = "months", maxview = "years", ...)
}
\arguments{
\item{inputId}{Definerer Namespace}

\item{label}{Visningstekst til brukerkontrollen}

\item{minview}{Fineste nivå tidsoppløsning}

\item{maxview}{Groveste nivå tidsoppløsning}
}
\value{
En brukerkontroll for bruk i R-shiny som tillater valg av f.eks. kun år og måned
       uten å angi dag.
}
\description{
R-shiny alternativ til dateInput med mulighet til  bare år, måned og år
}
