% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_fordelingsfig_prepost.R
\name{fordelingsfig_prepost_server}
\alias{fordelingsfig_prepost_server}
\title{Modul for før/etter-figur i NRA sin shiny-app på Rapporteket}
\usage{
fordelingsfig_prepost_server(
  id,
  reshID,
  RegData,
  hvd_session,
  userRole,
  BrValg
)
}
\arguments{
\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}
}
\value{
Serverdel av før/etter-figur
}
\description{
Kun til bruk i Shiny
}
