% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_fordelingsfig_prepost.R
\name{fordelingsfig_prepost_ui}
\alias{fordelingsfig_prepost_ui}
\title{Modul for før/etter-figur i NRA sin shiny-app på Rapporteket}
\usage{
fordelingsfig_prepost_ui(id)
}
\value{
UI-del av før/etter-figur
}
\description{
Kun til bruk i Shiny
}
