% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_fordelingsfig.R
\name{fordelingsfig_server}
\alias{fordelingsfig_server}
\title{Modul for fordelingsfigurer i NRA sin shiny-app på Rapporteket}
\usage{
fordelingsfig_server(id, reshID, RegData, hvd_session, BrValg)
}
\arguments{
\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}
}
\value{
Serverdelen av fordelingsfigurmodul
}
\description{
Kun til bruk i Shiny
}
