% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_fordelingsfig.R
\name{fordelingsfig_ui}
\alias{fordelingsfig_ui}
\title{Modul for fordelingsfigurer i NRA sin shiny-app på Rapporteket}
\usage{
fordelingsfig_ui(id)
}
\value{
ui-del av fordelingsfigurmodul
}
\description{
Kun til bruk i Shiny
}
