% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_gjsn_prepost.R
\name{gjsn_prepost_server}
\alias{gjsn_prepost_server}
\title{Modul for andeler før og etter i NRA sin shiny-app på Rapporteket}
\usage{
gjsn_prepost_server(id, reshID, RegData, hvd_session)
}
\arguments{
\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}
}
\value{
Serverdel av figur med andeler før og etter
}
\description{
Kun til bruk i Shiny
}
