% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_gjsn_prepost.R
\name{gjsn_prepost_ui}
\alias{gjsn_prepost_ui}
\title{Modul for andeler før og etter i NRA sin shiny-app på Rapporteket}
\usage{
gjsn_prepost_ui(id)
}
\value{
UI-del av figur med andeler før og etter
}
\description{
Kun til bruk i Shiny
}
