% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_indikatorfig.R
\name{indikatorfig_server}
\alias{indikatorfig_server}
\title{Modul for indikatorfigurer i NRA sin shiny-app på Rapporteket}
\usage{
indikatorfig_server(id, RegData, hvd_session)
}
\value{
Serverdel av indikatorfane
}
\description{
Kun til bruk i Shiny
}
