% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nraIndikatorMaalnivaa_basis.R
\name{indikatorFigRaterGrVar_basis}
\alias{indikatorFigRaterGrVar_basis}
\title{Plot andeler/andeler i angitt format}
\usage{
indikatorFigRaterGrVar_basis(
  RegData,
  valgtVar = "reduksjon_lekkasje50",
  outfile = "",
  width = 600,
  height = 600,
  decreasing = F,
  terskel = 0,
  minstekrav = NA,
  maal = NA,
  xtekst = "Andel \%",
  til100 = F,
  skriftStr = 1.3,
  pktStr = 1.5,
  datoFra = "2016-01-01",
  datoTil = "2050-12-31",
  hentData = F,
  preprosess = T,
  minald = 0,
  maxald = 130,
  erMann = "",
  forlopstype1 = 99,
  forlopstype2 = 99,
  onestage = 99
)
}
\value{
Et plot av andeler over tre år
}
\description{
Plot andeler/andeler i angitt format
}
