% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastShinyData.R
\name{lastShinyData}
\alias{lastShinyData}
\title{Last inn data for NRA sin shiny app}
\usage{
lastShinyData()
}
\value{
RegData En dataramme med registerets data
}
\description{
Forusetter tilgang til database
}
