% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nraBeregnIndikator.R
\name{nraBeregnIndikator}
\alias{nraBeregnIndikator}
\title{Denne funksjonen beregner NRAs indikatorer fordelt over år}
\usage{
nraBeregnIndikator(RegData, valgtVar)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel skal plottes}
}
\value{
En dataramme med valgt indikator
}
\description{
Filtreringer skjer i funksjonen
}
