% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nraFigIndikator.R
\name{nraFigIndikator}
\alias{nraFigIndikator}
\title{Gi en visuell fremstilling av registerets indikatorer over tid}
\usage{
nraFigIndikator(
  indikatordata,
  tittel = "",
  terskel = 30,
  minstekrav = NA,
  maal = NA,
  skriftStr = 1.3,
  pktStr = 1.4,
  legPlass = "top",
  minstekravTxt = "Min.",
  maalTxt = "Mål",
  graaUt = NA,
  decreasing = F,
  outfile = "",
  lavDG = NA,
  width = 800,
  height = 700,
  inkl_konf = F,
  maalretn = "hoy",
  desimal = FALSE,
  xmax = NA
)
}
\arguments{
\item{indikatordata}{En dataramme med følgende kolonner:
- ReshId
- Aar
- Teller
- SenterKortNavn}
}
\description{
Gi en visuell fremstilling av registerets indikatorer over tid
}
