% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nraFigIndikator_v3.R
\name{nraFigIndikator_v3}
\alias{nraFigIndikator_v3}
\title{Gi en visuell fremstilling av registerets indikatorer over tid}
\usage{
nraFigIndikator_v3(
  indikatordata,
  tittel = "",
  terskel = 30,
  minstekrav = NA,
  maal = NA,
  skriftStr = 1.3,
  pktStr = 1.4,
  legPlass = "top",
  minstekravTxt = "Min.",
  maalTxt = "Mål",
  graaUt = NA,
  decreasing = F,
  outfile = "",
  lavDG = NA,
  width = 800,
  height = 700,
  maalretn = "hoy",
  desimal = FALSE,
  xmax = NA,
  lavDGtekst = "Dekningsgrad < 60 \%"
)
}
\arguments{
\item{indikatordata}{En dataramme med følgende kolonner:
- AvdRESH
- year
- var
- SenterKortNavn}
}
\description{
Gi en visuell fremstilling av registerets indikatorer over tid
}
