% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nraIndikatorSfinkter.R
\name{nraIndikatorSfinkter}
\alias{nraIndikatorSfinkter}
\title{Denne lager en figur som er en sammenstilling av alle kvalitetsindikatorene
til NRA ved snm.}
\usage{
nraIndikatorSfinkter(
  outfile = "",
  dekngrad = 82,
  stmark9 = 41,
  stmarks12 = 50,
  width = 7.5,
  height = 4.5
)
}
\arguments{
\item{outfile}{Navn på fil figuren skrives til. Default: '' (Figur skrives
til systemets default output device (som regel skjerm))}
}
\value{
En figur med kvalitetsindikatorer med måloppnåelse
}
\description{
Denne lager en figur som er en sammenstilling av alle kvalitetsindikatorene
til NRA ved snm.
}
