% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nraIndikatorSNM.R
\name{nraIndikatorSNM}
\alias{nraIndikatorSNM}
\title{Denne lager en figur som er en sammenstilling av alle kvalitetsindikatorene
til NRA ved snm.}
\usage{
nraIndikatorSNM(outfile = "", dekngrad = 85, posTestBedr75 = 73,
  stmark9 = 30, stmarks12 = 44, infeksjon = 1.6)
}
\arguments{
\item{outfile}{Navn på fil figuren skrives til. Default: '' (Figur skrives
til systemets default output device (som regel skjerm))}
}
\value{
En figur med kvalitetsindikatorer med måloppnåelse
}
\description{
Denne lager en figur som er en sammenstilling av alle kvalitetsindikatorene
til NRA ved snm.
}
