% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nraFigAndeler.R
\name{nraPlotAndeler}
\alias{nraPlotAndeler}
\title{Plot andelsfigur}
\usage{
nraPlotAndeler(PlotParams, fargepalett = "BlaaRapp", outfile = "")
}
\arguments{
\item{outfile}{Navn på fil figuren skrives til. Default: '' (Figur skrives
til systemets default output device (som regel skjerm))}
}
\value{
Fordelingsfigur
}
\description{
Denne funksjonen plotter figur basert på beregninger i nraBeregnAndeler
}
