% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nraPreprosess.R
\name{nraPreprosess}
\alias{nraPreprosess}
\title{Denne funksjonen definerer en del nye (sammensatte) variabler relevante for rapporter i NRA
og rensker opp i andre.}
\usage{
nraPreprosess(RegData)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}
}
\value{
Data En list med det filtrerte datasettet og sykehusnavnet som tilsvarer reshID
}
\description{
Denne funksjonen definerer en del nye (sammensatte) variabler relevante for rapporter i NRA
og rensker opp i andre.
}
