% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nraSNMdagbokPst.R
\name{nraSNMdagbokPst}
\alias{nraSNMdagbokPst}
\title{Denne funksjonen oppsummerer SNM-dagboken med gjennomsnitt før og under test.
Bør den inkludere konfidensintervaller?}
\usage{
nraSNMdagbokPst(
  RegData,
  datoFra = "2012-04-01",
  datoTil = "2050-12-31",
  valgtShus = "",
  outfile = "",
  preprosess = TRUE,
  minald = 0,
  maxald = 130,
  enhetsUtvalg = 0,
  erMann = 99,
  reshID,
  hentData = F,
  forlopstype1 = 99,
  forlopstype2 = 99
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{valgtShus}{Vektor med AvdResh over hvilke sykehus man genererer rapporten for.
Denne overstyrer reshID og er bare tilgjengelig for SC-bruker.}

\item{outfile}{Navn på fil figuren skrives til. Default: '' (Figur skrives
til systemets default output device (som regel skjerm))}

\item{preprosess}{Preprosesser data
FALSE: Nei (Default)
TRUE: Ja}

\item{minald}{Alder, fra og med (Default: 0)}

\item{maxald}{Alder, til og med (Default: 130)}

\item{enhetsUtvalg}{Lag figur for
0: Hele landet
1: Egen enhet mot resten av landet (Default)
2: Egen enhet}

\item{erMann}{kjønn
1: menn
0: kvinner
99: begge (alt annet enn 0 og 1) (Default)}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{hentData}{Gjør spørring mot database
FALSE: Nei, RegData gis som input til funksjonen (Default)
TRUE: Ja}

\item{forlopstype1}{Type forløp
1: Sfinkterplastikk
2: SNM
3: Oppfølging 1 år
4: Oppfølging 3 år}

\item{forlopstype2}{Type SNM-forløp (til dels) avhengig av testkonklusjon
1: Test - En test det ikke konkluderes på: Forløpet avsluttes. Ofte vil ny test
   foretas men ev. nytt forløp har ingen kobling til dette forløpet
2: Test og eventuell implantasjon - Permanent implantasjon tilbys
3: Revisjon - Et eksisterende implantat revideres, ingen kobling til ev. opprinnelig forløp
4: Eksplantasjon Et eksisterende implantat eksplanteres, ingen kobling til ev. opprinnelig forløp
5: Test eksplantasjon -}
}
\value{
En figur med gjennomsnitt før operasjon, etter 1 år og etter 5 år
}
\description{
Denne funksjonen oppsummerer SNM-dagboken med gjennomsnitt før og under test.
Bør den inkludere konfidensintervaller?
}
