% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nraUtvalg.R
\name{nraUtvalg}
\alias{nraUtvalg}
\title{Funksjon som gjør utvalg av dataene, returnerer det reduserte datasettet og utvalgsteksten.}
\usage{
nraUtvalg(
  RegData,
  datoFra,
  datoTil,
  valgtShus = "",
  minald = 0,
  maxald = 130,
  erMann = 99,
  forlopstype1 = 99,
  forlopstype2 = 99,
  fargepalett = "BlaaRapp"
)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{valgtShus}{Vektor med AvdResh over hvilke sykehus man genererer rapporten for.
Denne overstyrer reshID og er bare tilgjengelig for SC-bruker.}

\item{minald}{Alder, fra og med (Default: 0)}

\item{maxald}{Alder, til og med (Default: 130)}

\item{erMann}{kjønn
1: menn
0: kvinner
99: begge (alt annet enn 0 og 1) (Default)}

\item{forlopstype1}{Type forløp
1: Sfinkterplastikk
2: SNM
3: Oppfølging 1 år
4: Oppfølging 3 år}

\item{forlopstype2}{Type SNM-forløp (til dels) avhengig av testkonklusjon
1: Test - En test det ikke konkluderes på: Forløpet avsluttes. Ofte vil ny test
   foretas men ev. nytt forløp har ingen kobling til dette forløpet
2: Test og eventuell implantasjon - Permanent implantasjon tilbys
3: Revisjon - Et eksisterende implantat revideres, ingen kobling til ev. opprinnelig forløp
4: Eksplantasjon Et eksisterende implantat eksplanteres, ingen kobling til ev. opprinnelig forløp
5: Test eksplantasjon -}

\item{fargepalett}{Hvilken fargepalett skal brukes i figurer (Default: BlaaRapp)}
}
\value{
UtData En liste bestående av det filtrerte datasettet, utvalgstekst for figur og tekststreng som angir fargepalett
}
\description{
Funksjon som gjør utvalg av dataene, returnerer det reduserte datasettet og utvalgsteksten.
}
