% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_startside.R
\name{startside_server}
\alias{startside_server}
\title{Server-modul for startsiden til NRA sin Rapporteket-app}
\usage{
startside_server(id, usrRole)
}
\description{
Server-modul for startsiden til NRA sin Rapporteket-app
}
