% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nraPrepVar.R
\name{nraPrepVar}
\alias{nraPrepVar}
\title{Preparer variabler for plotting}
\usage{
nraPrepVar(RegData, valgtVar)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel skal plottes}
}
\value{
PrepData En liste med plotrelevante størrelser
}
\description{
Denne funksjonen grupperer og klargjør variabler for andelsplot
}
\details{
Her kan detaljer skrives
}

